Exception handling - delete

+++ Curso/src/main/java/com/CursoJava_JPA_Hibernate/Curso/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.CursoJava_JPA_Hibernate.Curso.Repositories.UserRepository;
 import com.CursoJava_JPA_Hibernate.Curso.entities.User;
+import com.CursoJava_JPA_Hibernate.Curso.services.exceptions.DatabaseException;
 import com.CursoJava_JPA_Hibernate.Curso.services.exceptions.ResourceNotFoundException;
 
 /*As classes de serviço contém a lógica de négocio e apontam 
@@ -35,7 +37,16 @@ public class UserService {
     }
 
     public void deleteUser(Long id){
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }else{
+            try {
+                userRepository.deleteById(id);
+            } catch (DataIntegrityViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        }
+        
     }
 
     public User updateUser(Long id, User obj){
